requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-quill-editor==0.1.28
Pillow==8.4.0
six==1.16.0
sqlparse==0.4.2
